Bowercopy config and ZeroClipboard tweaks.
This change configures bower.json and bowercopy to copy ZeroClipboard
and tooltipster directly from the sources, and it updates tooltipster
to the latest version.  This allows us to drive tooltips
programmatically with the curent API, so we can force the tooltip
to show when you click on the "Copy" button.

It also fiddles with ZeroClipboard's odd CSS to get highlighting to
work when you hover over the button.