Improvments to the Share dialog.
Full Screen is not shown as an option for no-owner files, since
the URLs won't work.

Hyperlinks with tooltips are added explaining the three types of
sharing, and letting you go to them right away.

As secondary feedback when you click a Copy button, we also
select the URL text.