Instant feedback when the run button is pressed.
When you press the "run" button, it now immediately switches to
the "stop" button for one second, regardless of whether the program
is actually interruptable (yet).  The reason is to just give the
user instant feedback that their run is actually underway.

If the program is actually interruptable, a polling loop will come
in, cancel the one-second timer, and start maintaining the status of
the stop-button based on the actual run status of the program.

If the program is not interruptable (e.g. it's a plain HTML page
or a script that isn't IDE-aware), then the stop-button will
just flip back to a run button after one second.