Remove newlines from pencil code BODY element.
This change alters the boilerplate HTML wrapper for pencil
code programs so that newlines do not appear inside the body
text.  The reason is to allow exercises like this one:

http://jlhwths.pencilcode.net.dev/edit/pyramid

That exercise sets $('body').css { whiteSpace: 'pre' } in order to
old-fashioned simulate monospace formatting for text appended to
the body.  Without this change, flipping the body to 'pre' will
display a bunch of blank lines.  With this change, the body starts
nicely empty, with no extra newline or whitespace characters.