Fix "save" button disabling on nopass accounts.
No-password accounts used to never unhighlight the "save" button
even right after saving a file.  This is now fixed.

Here is the reason: since we want "save" to be highlighted when
the user is not logged in, we keep the button highlighted when
model.username (the name of the logged in user) is unset.

However, model.username would enver get set for no-password users
because they never entered a password, and username was set as
part of the password-entry flow.

Now if you successfully save without a password, we set username
and consider you logged in.  We also correctly save a logged-in-cookie
with your username (without a password) so that if you return to your
no-pass account on the same machine, you are still considered logged in.