Put paneboxes over panes.
By changing the top-level DOM structure of the panes this way, we
can allow UI from the panes to overflow out of the pane over to
the title, while still setting overflow:hidden on the panebox
containing both the pane and the panetitle.   We will use this
for adding extra UI attached to the droplet palette.